Fix Unicode (UTF-8) writing issues
This would only happen on Windows. I assumed Python defaulted to UTF-8
for everything, but that seems to not be the case.

There is also one player whose name is mangled by NCAA. I understand how
the transformation happens (ć is \u0107, \xC4 \x87 in UTF-8, and Ä is
\u00C4 etc), but I cannot come up with a good way to undo the transform
without doing it more than needed.